+++ src/Globber.cpp
@@ -41,7 +41,7 @@ Globber::Globber(std::vector<std::string> start_paths,
 		bool recurse_subdirs,
 		int dirjobs,
 		sync_queue<std::string>& out_queue)
-		: m_start_paths(start_paths.cbegin(), start_paths.cend()),
+		: m_start_paths(start_paths),
 		  m_type_manager(type_manager),
 		  m_dir_inc_manager(dir_inc_manager),
 		  m_recurse_subdirs(recurse_subdirs),
@@ -58,8 +58,6 @@ Globber::~Globber()
 
 void Globber::Run()
 {
-	set_thread_name("Globber");
-
 	char * dirs[m_start_paths.size()+1];
 
 	/// @todo It looks like OSX needs any trailing slashes to be removed here, or its fts lib will double them up.
@@ -107,7 +105,7 @@ void Globber::Run()
 	// Start the directory traversal threads.  They will all initially block on dir_queue, since it's empty.
 	for(int i=0; i<m_dirjobs; i++)
 	{
-		threads.push_back(std::thread(&Globber::RunSubdirScan, this, std::ref(dir_queue)));
+		threads.push_back(std::thread(&Globber::RunSubdirScan, this, std::ref(dir_queue), i));
 	}
 
 	LOG(INFO) << "Globber threads = " << threads.size();
@@ -129,24 +127,37 @@ void Globber::Run()
 		thr.join();
 	}
 
-	LOG(INFO) << "Number of regular files found: " << m_num_files_found;
+	///LOG(INFO) << "Number of regular files found: " << m_num_files_found;
 }
 
 
-void Globber::RunSubdirScan(sync_queue<std::string> &dir_queue)
+void Globber::RunSubdirScan(sync_queue<std::string> &dir_queue, int thread_index)
 {
 	char * dirs[2];
 	std::string dir;
+	// Local copy of a stats struct that we'll use to collect up stats just for this thread.
+	DirectoryTraversalStats stats;
+
+	// Set the name of the thread.
+	std::stringstream temp_ss;
+	temp_ss << "GLOBBER_";
+	temp_ss << thread_index;
+	set_thread_name(temp_ss.str());
 
 	while(dir_queue.wait_pull(std::move(dir)) != queue_op_status::closed)
 	{
-		/// The number of directories pushed onto the work queue by this thread during this iteration.
-		size_t num_dirs_pushed {0};
-
 		dirs[0] = const_cast<char*>(dir.c_str());
 		dirs[1] = 0;
 
-		FTS *fts = fts_open(dirs, FTS_LOGICAL  /*| FTS_NOSTAT*/, NULL);
+		/// @todo We can't use FS_NOSTAT here.  OSX at least isn't able to determine regular
+		/// files without the stat, so they get returned as FTS_NSOK / 11 /	no stat(2) requested.
+		/// Does not seem to affect performance on Linux, but might be having an effect on Cygwin.
+		/// Look into workarounds.
+		/// @note Per looking at the fts_open() source, FTS_LOGICAL turns on FTS_NOCHDIR, but since we're traversing
+		/// in multiple threads, and there's only a process-wide cwd, we'll specify it anyway.
+		/// @todo Current gnulib supports additional flags here: FTS_CWDFD | FTS_DEFER_STAT | FTS_NOATIME.  We should
+		/// check for these and use them if they exist.
+		FTS *fts = fts_open(dirs, FTS_LOGICAL | FTS_NOCHDIR /*| FTS_NOSTAT*/, NULL);
 		while(FTSENT *ftsent = fts_read(fts))
 		{
 			std::string name;
@@ -161,8 +172,11 @@ void Globber::RunSubdirScan(sync_queue<std::string> &dir_queue)
 			LOG(INFO) << "Considering file: " << ftsent->fts_path;
 			if(ftsent->fts_info == FTS_F)
 			{
+				// It's a normal file.
 				LOG(INFO) << "... normal file.";
-				// It's a normal file.  Check for inclusion.
+				stats.m_num_files_found++;
+
+				// Check for inclusion.
 				if(m_type_manager.FileShouldBeScanned(name))
 				{
 					LOG(INFO) << "... should be scanned.";
@@ -170,16 +184,22 @@ void Globber::RunSubdirScan(sync_queue<std::string> &dir_queue)
 					m_out_queue.wait_push(std::move(path));
 
 					// Count the number of files we found that were included in the search.
-					m_num_files_found++;
+					stats.m_num_files_scanned++;
+				}
+				else
+				{
+					stats.m_num_files_rejected++;
 				}
 			}
 			else if(ftsent->fts_info == FTS_D)
 			{
 				LOG(INFO) << "... directory.";
+				stats.m_num_directories_found++;
 				// It's a directory.  Check if we should descend into it.
 				if(!m_recurse_subdirs && ftsent->fts_level > FTS_ROOTLEVEL)
 				{
 					// We were told not to recurse into subdirectories.
+					LOG(INFO) << "... --no-recurse specified, skipping.";
 					fts_set(fts, ftsent, FTS_SKIP);
 				}
 				if(m_dir_inc_manager.DirShouldBeExcluded(path, name))
@@ -189,11 +209,23 @@ void Globber::RunSubdirScan(sync_queue<std::string> &dir_queue)
 					fts_set(fts, ftsent, FTS_SKIP);
 				}
 
-				if(m_recurse_subdirs && ftsent->fts_level > FTS_ROOTLEVEL)
+				if((m_dirjobs > 1) && (ftsent->fts_level == FTS_ROOTLEVEL))
+				{
+					// We're doing things multithreaded, so we have to detect cycles ourselves.
+					if(HasDirBeenVisited(dev_ino_pair(ftsent->fts_dev, ftsent->fts_ino).m_val))
+					{
+						// Found cycle.
+						WARN() << "\'" << ftsent->fts_path << "\': recursive directory loop";
+						fts_set(fts, ftsent, FTS_SKIP);
+						continue;
+					}
+				}
+
+				if(m_recurse_subdirs && ftsent->fts_level > FTS_ROOTLEVEL && m_dirjobs > 1)
 				{
 					// Queue it up for scanning.
+					LOG(INFO) << "... subdir, queuing it up for multithreaded scanning.";
 					dir_queue.wait_push(std::move(path));
-					num_dirs_pushed++;
 					fts_set(fts, ftsent, FTS_SKIP);
 				}
 			}
@@ -217,6 +249,16 @@ void Globber::RunSubdirScan(sync_queue<std::string> &dir_queue)
 				NOTICE() << "Could not get stat info at path \'" << ftsent->fts_path << "\': "
 									<< LOG_STRERROR(ftsent->fts_errno) << ". Skipping.";
 			}
+			else if(ftsent->fts_info == FTS_DC)
+			{
+				// Directory that causes cycles.
+				WARN() << "\'" << ftsent->fts_path << "\': recursive directory loop";
+			}
+			else if(ftsent->fts_info == FTS_SLNONE)
+			{
+				// Broken symlink.
+				WARN() << "Broken symlink: \'" << ftsent->fts_path << "\'";
+			}
 			else
 			{
 				LOG(INFO) << "... unknown file type:" << ftsent->fts_info;
@@ -224,4 +266,6 @@ void Globber::RunSubdirScan(sync_queue<std::string> &dir_queue)
 		}
 		fts_close(fts);
 	}
+
+	/// @todo Add the local stats to the class's stats.
 }