+++ src/Globber.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Gary R. Van Sickle (grvs@users.sourceforge.net).
+ * Copyright 2015-2016 Gary R. Van Sickle (grvs@users.sourceforge.net).
  *
  * This file is part of UniversalCodeGrep.
  *
@@ -23,16 +23,42 @@
 #include <config.h>
 
 #include <vector>
-#include <deque>
+#include <set>
 #include <string>
 #include <thread>
-#include <future>
+#include <cstdint>
+#include <sys/types.h> // for dev_t, ino_t
 
 #include "sync_queue_impl_selector.h"
 
 // Forward decls.
 class TypeManager;
 class DirInclusionManager;
+
+/**
+ * Helper struct to collect up and communicate traversal stats.
+ */
+struct DirectoryTraversalStats
+{
+	size_t m_num_files_found { 0 };
+
+	size_t m_num_files_rejected { 0 };
+
+	size_t m_num_files_scanned { 0 };
+
+	size_t m_num_directories_found { 0 };
+};
+
+// Boost has a template of this nature, but of course more complete.
+template <unsigned char NumBits>
+struct uint_t
+{
+	static_assert(NumBits <= 128, "NumBits > 128 not supported");
+	using fast = typename uint_t<NumBits+1>::fast;
+};
+template<> struct uint_t<128> { using fast = unsigned __int128; };
+template<> struct uint_t<64> { using fast = uint_fast64_t; };
+template<> struct uint_t<32> { using fast = uint_fast32_t; };
 
 /**
  * This class does the directory tree traversal.
@@ -56,9 +82,9 @@ public:
 
 private:
 
-	void RunSubdirScan(sync_queue<std::string> &dir_queue);
+	void RunSubdirScan(sync_queue<std::string> &dir_queue, int thread_index);
 
-	std::deque<std::string> m_start_paths;
+	std::vector<std::string> m_start_paths;
 
 	TypeManager &m_type_manager;
 
@@ -70,9 +96,22 @@ private:
 
 	sync_queue<std::string>& m_out_queue;
 
-	long m_num_files_found = {0};
+	using dev_ino_pair_type = uint_t<(sizeof(dev_t)+sizeof(ino_t))*8>::fast;
+	struct dev_ino_pair
+	{
+		dev_ino_pair(dev_t d, ino_t i) noexcept { m_val = d, m_val <<= sizeof(ino_t)*8, m_val |= i; };
+
+		dev_ino_pair_type m_val;
+	};
+
+	std::mutex m_dir_mutex;
+	std::set<dev_ino_pair_type> m_dir_has_been_visited;
+	bool HasDirBeenVisited(dev_ino_pair_type di) { std::unique_lock<std::mutex> lock(m_dir_mutex); return !m_dir_has_been_visited.insert(di).second; };
+
+	DirectoryTraversalStats m_traversal_stats;
 
 	std::string m_bad_path;
 };
+
 
 #endif /* GLOBBER_H_ */