+++ src/ResizableArray.h
@@ -30,9 +30,6 @@ inline void* aligned_alloc(size_t algn, size_t size) { void *p=0; posix_memalign
 #error "Could not find aligned memory allocator."
 #endif
 
-#include "Logger.h"
-
-
 /**
  * This is sort of a poor-man's std::allocator<>, without the std.  We use it in the File() constructor
  * to get an uninitialized buffer to read the file data into.  By instantiating one of these objects prior to a loop of
@@ -46,21 +43,9 @@ class ResizableArray
 	// Would pass this in as a defaulted template param, but gcc complains that "error: requested alignment is not an integer constant"
 	// no matter what I do (including initializing this var to the template param).
 	// Also can't make it a "static constexpr int alignment {32}", since that's broken on gcc ~4.8.4.
-#define alignment 64
+#define alignment 32
 
 public:
-
-	/// @todo alignas() isn't really correct in an alias, but it works under various versions of GCC.  Clang rejects it.
-	/// The attribute seems to be acceptable to everyone, though how well it works is an open question.
-	using element_type /*alignas(alignment)*/ __attribute__((aligned(alignment))) = T;
-
-	using pointer = element_type *;
-
-	using const_pointer = const element_type *;
-
-	/// @todo This static assert just doesn't work under clang 3.5.
-	//static_assert(alignof(decltype(*std::declval<const_pointer>())) == alignment, "alignment isn't working"); //__attribute__((aligned(alignment)))
-
 	ResizableArray() noexcept = default;
 	~ResizableArray() noexcept
 	{
@@ -70,7 +55,7 @@ public:
 		}
 	};
 
-	const_pointer data() const noexcept __attribute__((malloc)) { return m_current_buffer; };
+	T *__attribute__((aligned(alignment))) data() const noexcept __attribute__((malloc)) { return m_current_buffer; };
 
 	void reserve_no_copy(std::size_t needed_size)
 	{
@@ -83,14 +68,13 @@ public:
 			}
 
 			m_current_buffer_size = needed_size;
-			m_current_buffer = static_cast<pointer>(aligned_alloc(alignment, ((m_current_buffer_size*sizeof(element_type)+(alignment-1))/alignment)*alignment));
-			//LOG(INFO) << "filedata alignof: " << alignof(*m_current_buffer) << "," << alignof(element_type) << std::endl;
+			m_current_buffer = static_cast<T*>(aligned_alloc(alignment, ((m_current_buffer_size*sizeof(T)+(alignment-1))/alignment)*alignment));
 		}
 	}
 
 private:
 	std::size_t m_current_buffer_size { 0 };
-	pointer m_current_buffer { nullptr };
+	T *__attribute__((aligned(alignment))) m_current_buffer { nullptr };
 #undef alignment
 };