+++ NEWS.md
@@ -5,6 +5,7 @@
 ### New Features
 - Experimental support for OSX and PC-BSD.
 - Now compiles and links with either or both of libpcre and libpcre2, if available.  Defaults to using libpcre2 for matching.
+- Directory tree traversal now uses more than one thread (two by default).  Can be overridden with new "--dirjobs" command-line parameter.  Overall performance improvement on all platforms vs. 0.2.2 (e.g., ~25% on Fedora 23 with hot cache).
 
 ### Changed
 - Improved error reporting when directory traversal runs into problems.